refactor: modularize flame map_area and region_map submodules (#3878)
Relocate GameMapArea and RegionMapComponent stacks under flame/map_area/
and flame/region_map/ with public barrels and legacy re-export shims.
Add repo.app_flame_map_import_boundary gate and SPEC layout notes.
