+++ app/lib/features/game/flame/map_area/game_map_area_view.dart
