+++ app/lib/features/game/flame/map_area/map_area.dart
@@ -1,0 +1,6 @@
+export 'game_map_area.dart';
+export 'game_map_area_background.dart';
+export 'game_map_area_civilian_draft_projection.dart';
+export 'game_map_area_fleet_draft_projection.dart';
+export 'game_map_area_province_action_states.dart';
+export 'game_map_area_state_logic.dart';
