+++ app/lib/features/game/flame/region_map/region_map.dart
@@ -1,0 +1,4 @@
+export 'region_map_boundary_visibility.dart';
+export 'region_map_component.dart';
+export 'region_map_province_overlay_geometry.dart';
+export 'region_map_viewport_snapshot.dart';
