+++ app/lib/features/game/flame/region_map/region_map_component_render_political.dart
