+++ tool/ct_repo_lint_manifest.yaml
@@ -438,6 +438,13 @@ rules:
     runner: dart
     script: tool/check_app_e2e_helper_location.dart
 
+  - rule_id: repo.app_flame_map_import_boundary
+    group: architecture
+    title: flame map submodules must follow one-way import graph (map_area/region_map vs map_state)
+    spec: SPEC/program/repo-and-packages.md
+    runner: dart
+    script: tool/check_app_flame_map_import_boundary.dart
+
   - rule_id: repo.app_event_bus_decoupling
     group: architecture
     title: app/lib must decouple via AppEventBus / appNavigatorKey choke point (Refs #2626)
